+++ Products/src/main/java/com/mycompany/app/record/ProductResponse.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 public record ProductResponse(
         Integer productId,
         String name,
-        Float price,
+        Integer price,
         Integer quantity,
         String imageURL,
         String description) { }
